Move blurbs into their own module
